Refatoração: concern para validação de documento
