+++ app/models/local_census_record.rb
@@ -1,10 +1,14 @@
 class LocalCensusRecord < ApplicationRecord
+  include DocumentValidation
+
   before_validation :sanitize
   before_save :sanitize_attributes
 
   validates :document_number, presence: true
   validates :document_type, presence: true
-  validates :document_type, inclusion: { in: User.document_type_options }
+  validates :document_type, inclusion: {
+    in: User.document_type_options, allow_blank: true
+  }
   validates :date_of_birth, presence: true
   validates :postal_code, presence: true
   validates :document_number, uniqueness: true
