+++ app/models/user.rb
@@ -1,5 +1,6 @@
 class User < ApplicationRecord
   include Verification
+  include DocumentValidation
 
   DOCUMENT_TYPES = %w(cpf rnm).freeze
 
@@ -100,8 +101,6 @@ class User < ApplicationRecord
   validate :validate_username_length
 
   validate :min_user_age
-
-  validate :cpf_number, if: :local_document?
 
   validates :gender, presence: true, allow_nil: true
   validates :ethnicity, presence: true, allow_nil: true
@@ -465,10 +464,6 @@ class User < ApplicationRecord
     devise_mailer.send(notification, self, *args).deliver_later
   end
 
-  def local_document?
-    document_number && document_type == 'cpf'
-  end
-
   def foreigner_document?
     document_type == 'rnm'
   end
@@ -499,16 +494,6 @@ class User < ApplicationRecord
           :min_user_age,
           scope: 'activerecord.errors.models.user.attributes.date_of_birth',
           age: min_age
-        )
-        errors.add(:base, message)
-      end
-    end
-
-    def cpf_number
-      unless CPF.valid?(document_number)
-        message = I18n.t(
-          :invalid_number,
-          scope: 'activerecord.errors.models.user.attributes.cpf'
         )
         errors.add(:base, message)
       end
